Fix two missed warnings from prior warning-cleanup commit
- MaxSAT.cc:860: use Lit::x member instead of invalid int cast for printf
- Enc_Adder.cc:125,171: cast vec::size() to size_t in assert to fix
  sign-compare against std::vector::size_type

+++ code/encodings/Enc_Adder.cc
@@ -122,7 +122,7 @@ void Adder::adderTree (Solver *S, std::vector< std::queue< Lit > > & buckets, ve
 // xs and ys must have the same size
 
 void Adder::lessThanOrEqual (Solver *S, vec< Lit > & xs, std::vector< uint64_t > & ys) {
-  assert ( xs.size() == ys.size() );
+  assert ( (size_t)xs.size() == ys.size() );
   vec<Lit> clause;
   bool skip;
   for ( int i = 0; i < (int)xs.size(); ++i ) {
@@ -168,7 +168,7 @@ void Adder::lessThanOrEqual (Solver *S, vec< Lit > & xs, std::vector< uint64_t >
 }
 
 void Adder::lessThanOrEqualInc (Solver *S, vec< Lit > & xs, std::vector< uint64_t > & ys, vec<Lit>& assumptions) {
-  assert ( xs.size() == ys.size() );
+  assert ( (size_t)xs.size() == ys.size() );
   vec<Lit> clause;
   bool skip;
   for ( int i = 0; i < (int)xs.size(); ++i ) {
